requirements.txt
openai
fastapi
black
isort
websockets
pydantic
langchain==0.0.105
uvicorn
jinja2
faiss-cpu
bs4
unstructured
libmagic
pypdf
python-dotenv
